Move persisting creds from view layer into DBMgr
This also refactors the common code to persist creds entered by user
during all op tasks that operate on a remtoe into a single method
to keep things DRY.
