+++ src/main/java/me/sheimi/sgit/database/models/Repo.java
@@ -601,4 +601,8 @@ public class Repo implements Comparable<Repo>, Serializable {
         } catch (StopTaskException e) {
         }
     }
+
+    public void saveCredentials() {
+        RepoDbManager.persistCredentials(getID(), getUsername(), getPassword());
+    }
 }
