+++ src/main/java/me/sheimi/sgit/repo/tasks/repo/RepoRemoteOpTask.java
@@ -1,0 +1,31 @@
+package me.sheimi.sgit.repo.tasks.repo;
+
+import me.sheimi.android.activities.SheimiFragmentActivity;
+import me.sheimi.sgit.database.models.Repo;
+
+/**
+ * Super class for Tasks that operate on a git remote
+ */
+
+public abstract class RepoRemoteOpTask extends RepoOpTask implements SheimiFragmentActivity.OnPasswordEntered {
+
+
+    public RepoRemoteOpTask(Repo repo) {
+        super(repo);
+    }
+
+
+    @Override
+    public void onClicked(String username, String password, boolean savePassword) {
+        mRepo.setUsername(username);
+        mRepo.setPassword(password);
+        if (savePassword) {
+            mRepo.saveCredentials();
+        }
+    }
+
+    @Override
+    public void onCanceled() {
+
+    }
+}
